chore: refactor out block explorer
